introduces envctl; shapefiles as zip
