+++ src/parcelgen/proc.py
@@ -1,4 +1,4 @@
-""" ParcelGen processing module
+"""Processing module
 """
 
 import json
